:art: Create PoseChecker class
